Change setting types to be resource
Because it makes sense to want to override settings
per-resource-per-language. Eg. you want to use Debug.Log only when
working on C# in Unity projects but not in other projects or other
languages within the same project.

VSCode will complain that these settings are not language-overridable
but will still allow it to be overriden per-language for any resource.