Log ID no longer used if it matches an item key
Streamlined the way log id and item keys are used for logging and how
they're handled during rotation.

The log ID is now a 'potential' log id and no colon is printed at the
end of it. It is not outputted if a log item key is more suitable.

This means things like `{a, b} = ...` now don't output the log id at all
but it starts with item keys right away.

Added tests for this case.